crop-long-picture/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "tt", email = "**"},
]
dependencies = [
    "pillow>=10.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

